% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStateOverview.R
\name{CreateOverview}
\alias{CreateOverview}
\title{Create overview from results}
\usage{
CreateOverview(lemma.set, writedir = NULL)
}
\arguments{
\item{lemma.set}{list of results from running \code{\link[LEMMA.forecasts]{RunOneCounty}} on multiple counties}

\item{writedir}{a character string giving a directory to write to}
}
\description{
Create overview from results
}
