% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateEmail.R
\name{GenerateEmail}
\alias{GenerateEmail}
\title{Generate email from results}
\usage{
GenerateEmail(max.date)
}
\arguments{
\item{max.date}{maximum date of data}
}
\description{
Generate email from results
}
