% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCountyData.R
\name{GetAdmits}
\alias{GetAdmits}
\title{Get information on hospital admissions}
\usage{
GetAdmits(remote = FALSE)
}
\arguments{
\item{remote}{a logical value, if \code{TRUE} download all data from remotes, otherwise use local data}
}
\description{
Get information on hospital admissions
}
