% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCountyInputs.R
\name{GetCountyInputs}
\alias{GetCountyInputs}
\title{Get input data for county}
\usage{
GetCountyInputs(county1, county.dt, doses.dt, remote = FALSE)
}
\arguments{
\item{county1}{a character string giving the name of the county}

\item{county.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetCountyData}}}

\item{doses.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetDosesData}}}

\item{remote}{a logical value, if \code{TRUE} download all data from remotes, otherwise use local data}
}
\description{
Get input data for county
}
