% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDosesData.R
\name{GetDosesData}
\alias{GetDosesData}
\title{Get doses data}
\usage{
GetDosesData(states = FALSE, remote = FALSE)
}
\arguments{
\item{remote}{a logical value, if \code{TRUE} download all data from remotes, otherwise use local data}
}
\description{
Get doses data
}
