% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCountyData_Scenario.R
\name{GetResults_scen}
\alias{GetResults_scen}
\title{Get results of running a scenario}
\usage{
GetResults_scen(projection, name)
}
\arguments{
\item{projection}{output of a LEMMA model run}

\item{name}{a character string giving the type of scenario to run}
}
\description{
This function is called from \code{\link[LEMMA.forecasts]{Scenario}}.
}
