% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCountyInputs.R
\name{ModifyCountyInputs}
\alias{ModifyCountyInputs}
\title{Get input data for county}
\usage{
ModifyCountyInputs(county1, inputs)
}
\arguments{
\item{county1}{a character string giving the name of the county}

\item{inputs}{a named list returned from \code{\link[LEMMA]{ProcessSheets}}}
}
\description{
Get input data for county
}
