% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDosesData.R
\name{ReadSFDoses}
\alias{ReadSFDoses}
\title{Read SF doses from AWS}
\usage{
ReadSFDoses(print_outdate = FALSE)
}
\arguments{
\item{print_outdate}{write dose number to console or not}
}
\description{
This function requires the following environment variables to be set:
"AWS_ACCESS_KEY_ID","AWS_SECRET_ACCESS_KEY", and "AWS_DEFAULT_REGION" (see \code{\link{EnvVar}}).
}
