% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStateOverview.R
\name{RtMap}
\alias{RtMap}
\title{Create Rt map results}
\usage{
RtMap(max.date, county.rt, writedir = NULL)
}
\arguments{
\item{max.date}{write me!}

\item{county.rt}{write me!}
}
\description{
write me!
}
