% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCounty_Scenario.R
\name{RunOneCounty_scen}
\alias{RunOneCounty_scen}
\title{Run a county scenario}
\usage{
RunOneCounty_scen(
  county1,
  county.dt,
  doses.dt,
  remote = FALSE,
  writedir = NULL
)
}
\arguments{
\item{county1}{a character string giving the name of the county}

\item{county.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetCountyData}}}

\item{doses.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetDosesData}}}

\item{remote}{a logical value, if \code{TRUE} download all data from remotes, otherwise use local data}

\item{writedir}{a character string giving a directory to write to, it should only be used if \code{remote} is \code{TRUE}.
This assumes the directory whose path is given already exists.}
}
\description{
Run a county scenario
}
