% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCounty_Scenario.R
\name{RunOneCounty_scen_input}
\alias{RunOneCounty_scen_input}
\title{Run a county scenario (adjustable for Shiny interface)}
\usage{
RunOneCounty_scen_input(
  county1,
  county.dt,
  doses.dt,
  k_uptake = "low",
  k_ukgrowth = 1,
  k_brgrowth = 1,
  k_max_open = 0.75,
  remote = FALSE,
  writedir = NULL
)
}
\arguments{
\item{county1}{a character string giving the name of the county}

\item{county.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetCountyData}}}

\item{doses.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetDosesData}}}

\item{k_uptake}{a character string, "low" or "high" giving vaccine uptake}

\item{k_ukgrowth}{growth rate of UK variant}

\item{k_brgrowth}{growth rate of BR variant}

\item{k_max_open}{percentage of pre-pandemic activity after reopening (scales contact rate)}

\item{remote}{a logical value, if \code{TRUE} download all data from remotes, otherwise use local data}

\item{writedir}{a character string giving a directory to write to, it should only be used if \code{remote} is \code{TRUE}.
This assumes the directory whose path is given already exists.}
}
\description{
Run a county scenario (adjustable for Shiny interface)
}
