% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCountyData_Scenario.R
\name{GetCountyInputs_scen}
\alias{GetCountyInputs_scen}
\title{Get county inputs for scenarios}
\usage{
GetCountyInputs_scen(
  county1,
  county.dt,
  doses.dt,
  k_uptake,
  k_ukgrowth,
  k_brgrowth,
  remote = FALSE,
  writedir = writedir
)
}
\arguments{
\item{county1}{a character string giving the name of the county}

\item{county.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetCountyData}}}

\item{doses.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetDosesData}}}

\item{k_uptake}{a character string, "low" or "high" giving vaccine uptake}

\item{k_ukgrowth}{growth rate of UK variant}

\item{k_brgrowth}{growth rate of BR variant}

\item{remote}{a logical value, if \code{TRUE} download all data from remotes, otherwise use local data}

\item{writedir}{a character string giving a directory to write to. Ignored if \code{remote} is \code{FALSE}}
}
\value{
a named list of values
}
\description{
This function is called from \code{\link[LEMMA.forecasts]{Scenario}}.
}
