% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCountyData_Scenario.R
\name{GetCountyInputs_scen}
\alias{GetCountyInputs_scen}
\title{Get county inputs for scenarios}
\usage{
GetCountyInputs_scen(
  county1,
  county.dt,
  doses.dt,
  k_ukgrowth,
  k_brgrowth,
  k_ingrowth,
  vaccine_uptake = NULL,
  vaccine_dosing = NULL,
  remote = FALSE,
  writedir = NULL
)
}
\arguments{
\item{county1}{a character string giving the name of the county}

\item{county.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetCountyData}}}

\item{doses.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetDosesData}}}

\item{k_ukgrowth}{growth rate of UK variant (B.1.1.7)}

\item{k_brgrowth}{growth rate of BR variant (P.1)}

\item{k_ingrowth}{growth rate of IN variant (B.1.617.2)}

\item{vaccine_uptake}{a numeric vector with 3 values, for vaccine uptake in age groups 12-15, 16-64, and 65+}

\item{vaccine_dosing}{a named list that requires specific input, see section \code{vaccine_dosing}, or \code{NULL} for no adjustment of doses available}

\item{remote}{a logical value, if \code{TRUE} download all data from remotes, otherwise use local data}

\item{writedir}{a character string giving a directory to write to, it should only be used if \code{remote} is \code{TRUE}.
This assumes the directory whose path is given already exists.}
}
\value{
a named list of values
}
\description{
This function is called from \code{\link[LEMMA.forecasts]{Scenario}}.
}
\section{vaccine_dosing}{

This is a named list that requires the following elements
\itemize{
 \item{vaccine_dosing_jj}{daily increase in J&J vaccine delivery}
 \item{vaccine_dosing_mrna}{daily increase in mRNA vaccines delivery}
 \item{vaccine_dosing_mrna_max}{increase over baseline in maximum number of doses per day for mRNA vaccines}
 \item{vaccine_dosing_jj_max}{increase over baseline in maximum number of doses per day for J&J vaccines}
}
}

