% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStateOverview.R
\name{RtMap}
\alias{RtMap}
\title{Create Rt map results}
\usage{
RtMap(max.date, county.rt, writedir = NULL)
}
\arguments{
\item{max.date}{maximum date of data}

\item{county.rt}{`data.table` object produced by \code{\link[LEMMA.forecasts]{CreateOverview}}}

\item{writedir}{a character string giving a directory to write to}
}
\description{
Create Rt map results
}
