% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCountyData_Scenario.R
\name{Scenario}
\alias{Scenario}
\title{Run a scenario}
\usage{
Scenario(
  filestr1,
  county1,
  county.dt,
  doses.dt,
  k_mu_beta_inter = NULL,
  lemma_statusquo = NULL,
  k_uptake = "low",
  k_ukgrowth = 1,
  k_brgrowth = 1,
  k_max_open = 0.75,
  remote = FALSE,
  writedir = NULL
)
}
\arguments{
\item{filestr1}{a character string giving the type of scenario to run}

\item{county1}{a character string giving the name of the county}

\item{county.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetCountyData}}}

\item{doses.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetDosesData}}}

\item{lemma_statusquo}{either \code{NULL} or the result of a call to \code{Scenario} with \code{filestr1 = "statusquo"}}

\item{k_uptake}{a character string, "low" or "high" giving vaccine uptake}

\item{k_ukgrowth}{growth rate of UK variant}

\item{k_brgrowth}{growth rate of BR variant}

\item{remote}{a logical value, if \code{TRUE} download all data from remotes, otherwise use local data}

\item{writedir}{a character string giving a directory to write to. Ignored if \code{remote} is \code{FALSE}}
}
\value{
a named list of values
}
\description{
This function is called from \code{\link[LEMMA.forecasts]{RunOneCounty_scen}}.
}
