% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCounty_Scenario.R
\name{Scenario}
\alias{Scenario}
\title{Run a scenario}
\usage{
Scenario(
  filestr1,
  county1,
  county.dt,
  doses.dt,
  k_mu_beta_inter = NULL,
  lemma_statusquo = NULL,
  k_ukgrowth = 1,
  k_brgrowth = 1,
  k_ingrowth = 1,
  k_max_open = 0.75,
  vaccine_uptake = NULL,
  vaccine_dosing = NULL,
  remote = FALSE,
  writedir = NULL
)
}
\arguments{
\item{filestr1}{a character string giving the type of scenario to run}

\item{county1}{a character string giving the name of the county}

\item{county.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetCountyData}}}

\item{doses.dt}{a \code{\link[data.table]{data.table}} object returned from \code{\link[LEMMA.forecasts]{GetDosesData}}}

\item{k_mu_beta_inter}{multiplier to contact rate required to get to 100\% reopening}

\item{lemma_statusquo}{either \code{NULL} or the result of a call to \code{Scenario} with \code{filestr1 = "statusquo"}}

\item{k_ukgrowth}{growth rate of UK variant (B.1.1.7)}

\item{k_brgrowth}{growth rate of BR variant (P.1)}

\item{k_ingrowth}{growth rate of IN variant (B.1.617.2)}

\item{k_max_open}{percentage of pre-pandemic activity after reopening (scales contact rate)}

\item{vaccine_uptake}{a numeric vector with 3 values, for vaccine uptake in age groups 12-15, 16-64, and 65+}

\item{vaccine_dosing}{a named list that requires specific input, see section \code{vaccine_dosing}, or \code{NULL} for no adjustment of doses available}

\item{remote}{a logical value, if \code{TRUE} download all data from remotes, otherwise use local data}

\item{writedir}{a character string giving a directory to write to, it should only be used if \code{remote} is \code{TRUE}.
This assumes the directory whose path is given already exists.}
}
\value{
a named list of values
}
\description{
This function is called from \code{\link[LEMMA.forecasts]{RunOneCounty_scen}}.
}
\section{vaccine_dosing}{

This is a named list that requires the following elements
\itemize{
 \item{vaccine_dosing_jj}{daily increase in J&J vaccine delivery}
 \item{vaccine_dosing_mrna}{daily increase in mRNA vaccines delivery}
 \item{vaccine_dosing_mrna_max}{increase over baseline in maximum number of doses per day for mRNA vaccines}
 \item{vaccine_dosing_jj_max}{increase over baseline in maximum number of doses per day for J&J vaccines}
}
}

